Move UnderlyingSink to separate file
